requirements.txt
django==3.0.6
djangorestframework
PyPDF2
openai